backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
opencv-python==4.8.1.78
ultralytics==8.0.200
numpy==1.24.4
Pillow==10.0.1
django-cors-headers==4.3.1